+++ app/Abstracts/AbstractBrokerAuthentication.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Abstracts;
+
+use App\Contracts\BrokerAuthenticationContract;
+
+abstract class AbstractBrokerAuthentication implements BrokerAuthenticationContract
+{
+}