+++ app/Contracts/BrokerContract.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Contracts;
+
+interface BrokerContract
+{
+    public function __construct(BrokerAuthenticationContract $authenticator, BrokerCommandContract $command);
+
+    public function getAuthenticator(): BrokerAuthenticationContract;
+
+    public function getCommand(): BrokerCommandContract;
+}