+++ app/Services/Brokers/xStation/Broker.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Services\Brokers\xStation;
+
+use App\Abstracts\AbstractBroker;
+
+readonly class Broker extends AbstractBroker
+{
+}