+++ app/Services/Brokers/xStation/BrokerAuthentication.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Services\Brokers\xStation;
+
+use App\Abstracts\AbstractBrokerAuthentication;
+
+class BrokerAuthentication extends AbstractBrokerAuthentication
+{
+}