+++ app/Services/Brokers/xStation/BrokerCommand.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Services\Brokers\xStation;
+
+use App\Abstracts\AbstractBrokerCommand;
+
+class BrokerCommand extends AbstractBrokerCommand
+{
+}